downloadData function reports if it did update a file (#83)
* downloadData function reports if it did update a file (rc=1) or not (rc=0)

+++ bin/fetch_clamspam.sh
@@ -73,11 +73,13 @@ fi
 
 . $SRCDIR/lib/updates/download_files.sh
 
-downloadDatas "$VARDIR/spool/clamspam/" "clamspam3" $randomize "clamav" "\|local_whitelist.ign2" "noexit"
+ret=$(downloadDatas "$VARDIR/spool/clamspam/" "clamspam3" $randomize "clamav" "\|local_whitelist.ign2" "noexit")
 
 ## restart clamspam daemon
-kill -USR2 `cat $VARDIR/run/clamav/clamspamd.pid 2>/dev/null` > /dev/null 2>&1
-log "Clamspam - Database reloaded"
+if [[ "$ret" -eq "1" ]]; then
+	kill -USR2 `cat $VARDIR/run/clamav/clamspamd.pid 2>/dev/null` > /dev/null 2>&1
+	log "Clamspam - Database reloaded"
+fi
 
 removeLockFile "$FILE_NAME"
